Create c code
Write a program that inputs N (assume N<30), and N integers into a 1-Dimensional array and an
integer value x. Your program should send these inputs to a function as parameters which should then
search for value x in the array and delete it from the array, displaying the updated array. Use function
find() that receives an array and an integer, it returns a new array that stores only the elements which
are not equal to x.